+++ Sources/BFKit/Linux/Foundation/StringExtension.swift
@@ -643,6 +643,7 @@ public extension String {
     #endif
 }
 
+/// Infix operator `???` with NilCoalescingPrecedence.
 infix operator ???: NilCoalescingPrecedence
 
 /// Returns defaultValue if optional is nil, otherwise returns optional.