+++ Tests/BFKitTests/Apple/UIKit/UIWindowExtensionTests.swift
@@ -81,4 +81,11 @@ class UIWindowExtensionTests: XCTestCase {
     func testDeactivateTouch() {
         window.deactivateTouch()
     }
+    
+    func testTouchImage() {
+        let image = window.touchImage()
+        
+        XCTAssert(image.size.height > 58 && image.size.height < 59)
+        XCTAssert(image.size.width > 58 && image.size.width < 59)
+    }
 }