+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.<br>
 ---
 
 ### 2.x Releases
+- `2.6.x` Releases - [2.6.0](#260---macos-support)
 - `2.5.x` Releases - [2.5.0](#250---better-gradients-and-dates)
 - `2.4.x` Releases - [2.4.0](#240---watchos-support) | [2.4.1](#241---swift-package-manager-returns)
 - `2.3.x` Releases - [2.3.0](#230---minor-breaking-changes)
@@ -49,13 +50,34 @@ All notable changes to this project will be documented in this file.<br>
 
 ## Master
 ### Added
-- Added `flatten<T>(_ array:)` generic global function in Array extension
+- Nothing
 
 ### Improved
 - Nothing
 
 ### Fixed
 - Nothing
+
+---
+
+## [2.6.0](https://github.com/FabrizioBrancati/BFKit-Swift/releases/tag/2.6.0) - macOS Support
+### 26 Jul 2017
+### Added
+- Added macOS Support
+- Added `dateSeparator` and `timeSeparator` to `init?(date:, time:)` with a `default` value, so no changes are required
+- Added `flatten<T>(_ array:)` generic global function in Array extension
+
+### Improved
+- Removed Apple Watch device from UIDevice extension (Can't be used)
+- Removed support for iOS 7 in UIImage extension (Refuse)
+- Improved BFPassword struct
+- Improved tests
+
+### Fixed
+- Fixed missing style in `init(frame:, style:, cellSeparatorStyle:, separatorInset:, dataSource:, delegate:)` in UITableView extension
+- Fixed `animate(path:, count: = default, duration:, autoreverses: = default)` autoreverses
+- Fixed BFButton with a default `fadeDuration` value
+- Fixed wrong default duration in `fade(label:, duration:, toAttributedString:)` in WKInterfaceController extension
 
 ---